Add missing bit and exit reasons to vmx.h
This patch is the first in a series of patches enabling the "Exit-Less
Interrupts" (ELI) feature for KVM on VMX. While ordinarily VMX causes
an exit for every physical interrupt, ELI allows KVM to determine that
certain interrupts should be handled directly in the guest, without exit.
This can significantly improve device-assignment I/O performance, as
the only remaining causes of exits - the interrupts and their completion -
are elliminated. The ELI technique is explained and evaluated in more detail
in the ASPLOS 2012 paper "ELI: Bare-Metal Performance for I/O Virtualization".

ELI needs to enable the "Descriptor-table exiting" bit of the VM-execution
controls, which will causes exits when the guest attempts to change the
IDT pointer through LIDT (or other changes to other descriptor tables).

In this patch, we give a name in vmx.h to this bit, and to the two exit
reasons which it leads to.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>