Shadow IDT initialization
With ELI, the guest is run using a "shadow IDT" instead of the guest's
requested IDT. This shadow IDT is built by KVM in a way that causes exits
for some interrupts while running the guest's normal handlers on others.

Unfortunately, the processor requires that the IDT location be given using a
virtual address, so we need to cause the guest to map a spare page which can
be used to hold the shadow IDT.

This patch allows a guest to specify, via a new hypercall, a location for
the shadow IDT in its own virtual memory space. When a guest calls this
hypercall, ELI is initialized for this guest, and the shadow IDT is located
in the given guest virtual address (GVA).

Note that this only initializes ELI, but does *not* enable it. To enable
ELI, it must be specified which physical interrupts should be assigned
to the guest (arrive directly in the guest, without exit), and of course
the shadow IDT needs to be built. This will be done in more patches below.

IMPORTANT NOTE: In this implementation we offer no protection against a
guest modifying the shadow IDT directly: The guest knows the GVA and GPA
of the shadow IDT and can modify it maliciously in one of two ways - either
by modifying its contents (writing to the page itself), or by changing its
page tables (assuming EPT is being used) to make the same GVA point to a
different GPA with different content. The ELI paper explains how both holes
can be avoided - the first by trapping writes to the shadow IDT, and the
second by using IDTR limit or by periodically verifying that the guest
hasn't change the mapping - but none of these solutions is built into
this patch set. Instead, it is recommended to either limit ELI's use to
non-malicious guests, or make this a non-issue by pinning critical host
interrupts to a core not running guests so that even malicous guests cannot
cause these critical interrupts to be lost.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>