Remap host-vector to guest-vector
In device assignment, the guest believes it configures the device on which
interrupt vectors to generate. However, what really happens is that the host
chooses different vectors, ones that are available in the host, and when the
host recieves the *host* vector, it injects (forwards) the different *guest*
vector into the guest.

With ELI, the interrupt will arrive directly at the guest, but the interrupt
vector received will be the *host* vector. So we need the shadow IDT to
contain, in position host-vector, the handler that the guest IDT set for
guest-vector.

This patch provides a function, kvm_arch_eli_remap_vector(), which KVM's
device assignment code will call (see the next patch) to remap a given
host-chosen IRQ to the guest-chosen vector. The host IRQ is specified,
not the vector, because device assignment only choses the IRQ and the
actual vector will only be chosen by the kernel later.

If ELI is not yet enabled, kvm_arch_eli_remap_vector() only remembers the
mapping but doesn't actually create the shadow IDT. This will be done when
the eli_remap() function, also provided in this patch, is called after
enabling ELI.

Note ELI assumes kvm device-assignment implementation. VFIO is not supported
by this patch.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>