When not in injection mode, trap NP exceptions
When in ELI but not in injection mode (defined in the previous patch),
all interrupts arrive at the guest, but the ones not assigned to the
guest cause an NP exception (because a handler for it is not present in
the shadow IDT).

In this patch, when needed we have these NP exceptions cause exits, so
the host can handle the non-assigned interrupts. Note that NP exceptions
completely unrelated to ELI are also possible, and they will cause
unnecessary exits, but the number of these is usually negligable.

In the next patch, we will handle these NP exceptions.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>