Handle NP exceptions
In the previous patch, we saw that in ELI, outside injection mode,
non-assigned interrupts cause an exit with an NP exception (while assigned
interrupts do not cause an exit at all).

In this patch, we handle these NP exceptions: We verify that it was really
caused by an external interrupt, and if it was we re-generate the same
interrupt, causing the host (Linux) to handle this interrupt as usual.

To regenerate the interrupt that caused the NP exception exit, we
use INT X instruction (software interrut)

Alternatively, it is also possible to send a self-IPI to regenerate
the interrupt.

ELI changes the way KVM handles interrupt completion (EOIs). With ELI
enabled, assigned interrupts could be delivered during guest mode execution
and might remain in service (pending EOI) after an exit. Thus, we always
acknowledge (EOI/ack_APIC_irq) on exit to complete pending interrupts and
let the hardware continue raising new interrupts.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>