EOI handling
As described in a previous patch, ELI switches to injection mode every time
KVM needs to inject a virtual interrupt. We need to switch back to the
regular ELI mode once we finish the injection. Thus, right after the guest
acknowledges a virtual interrupt that was delivered in injection mode, we
switch back to ELI mode.

Outside injection mode, i.e., on acknowledgment of assigned interrupt, we
have nothing special to do (we EOI on any exit anyway), and we just need
to circumvent the regular EOI handling of KVM. Moreover we'll show later
a mechanism (x2APIC) for avoiding exits on EOI in this case.

IMPORTANT NOTE: PV-EOI must be disabled (-cpu -kvm_pv_eoi) otherwise ELI will
remain in injection mode for longer periods of time causing external
interrupt exits for any physical interrupt.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>