Handle descriptor-table exits
ELI needs to trap guest changes to the IDTR, to prevent the guest replacing
the Shadow IDT enforced by the host. To do this, it must enable the
"Descriptor-table exiting" bit of the VM-execution controls. Unfortunately,
this bit enables exiting not only on IDTR changes, but also on changes of
the unrelated LDTR, GDTR and TR registers, and when any of these exits come,
we need to handle them appropriately.

The appropriate way to handle IDTR changes is to rebuild the shadow IDT
based on the instruction's operand, and the appropriate way to handle LDTR,
GDTR and TR exits is to turn off descriptor-table exiting (i.e., go to
injection mode), and run the guest for a while until it exits - letting it
handle these instructions normally (which is fine - KVM doesn't normally
need to trap and emulate these instructions).

In this version, however, we didn't do all of this yet. LDTR and
TR changes are handled correctly (via injection mode), but for GDTR and
IDTR (which share the same exit reason, so separating them requires further
decoding), we just disable ELI which is a valid, although drastic, solution.
We note that Linux guests rarely change any of these registers after boot,
so provided that ELI is only enabled after boot (by a command doing the
hypercall enabling ELI), this case will rarely happen anyway, and if it
does, a warning message can be noticed.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>