Add support for disabling halt exits
If we dedicate a physical core per VCPU (to improve performance), there
is no reason to force an exit when the guest executes HLT because the host
will not schedule other VCPU in the same core. HLT exits increase
the time the CPU spends in root mode (host context) and thus increase
the chances that an assigned physical interrupt will arrive in root mode
and not in guest mode. To achieve maximum performance, we strive to maximize the
number of physical assigned interrupts delivered in guest mode so better to
avoid HLT exits if we dedicate a core per VCPU.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>