Reserve vector for paravirtual posted interrupts
This patch is the first in a series of patches enabling exit-less
virtual interrupt injection, by implementing paravirtual Posted-Interrupts
using ELI.

This feature allows the KVM running one one core, to inject an interrupt to
a guest running on another core, without the guest exiting. This is done
by choosing one interrupt vector (the "posted interrupts" PI vector) and a
memory location. The host writes a vector to be injected to this memory
location, and sends the PI vector with an IPI to the core running the guest.
This guest is run with ELI to have this specific PI vector delivered
directly at the guest, without causing an exit. However, now we need some
cooperation at the guest (hence we call this feature paravirtual posted
interrupts):
The interrupt the guest received is the fixed PI vector, but upon getting
it, it needs to read the agreed memory location, and run the handler for
the vector written there.

This first patch of the series reservers a vector to send Exitless IPIs.
It also adds a new counter to KVM_STATS which shows how many virtual
interrupt injections were done using Exitless IPIs. If we send an Exitless
IPI to a running vcpu, in rare cases, the IPI might arrive in root mode
(host) because an exit ocurred while the IPI was being sent. Thus, we also
add a handler in the host for the IPI and a counter in /proc/interrupts

Signed-off-by: Abel Gordon <abelg@il.ibm.com>