Re-queue posted interrupts received in root mode
Paravirtual posted interrupts allows injecting virtual interrupts into a
running guest (without causing it to exiting to root mode first).
However, in rare cases, at the time we send a posted interrupt from core A
the guest running in core B may exit for an unrelated reason, and the IPI
we sent (from core A to core B) to signal the posted interrupt
arrives in the host, instead of the guest. In a previous patch we already
reserved this vector in the host. This patch introduces the handler, which
re-queues the injection using the traditional KVM injection mechanism
(because the guest is not running anymore).

Signed-off-by: Abel Gordon <abelg@il.ibm.com>