Initialize paravirtual posted interrupts support
The patch adds a hypercall for the guest to pass a pointer to a notification
vector variable and an injection page, and initialize paravirtual posted
interrupts support for all vcpus of the this guest.
It does not yet enable exitless injection with PV posted interrupts (this
will be done in the next patch).

The notification vector variable is used by the host to tell the guest
which vector number will be used to deliver posted interrupts (set
once during the initialization phase).
The injection page is used by the host to tell the guest which virtual
interrupt KVM wishes to inject each time a posted interrupt is sent.
(set every time a virtual interrupt is delivered using the paravirtual
posted interrupts mechanism).

Once the guests receives an interrupt that corresponds to the
number stored in the notification vector variable, the guests reads
the injections page to identify the virtual interrupt the host is
asking to inject and call the corresponding handler.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>