VMX implementation of PV posted interrupts
This patch implements the two kvm_x86_ops introduces above:
send_posted_interrupt and has_posted_interrupts

Every time KVM wishes  to inject a virtual interrupt to a given VCPU,
we check if the VCPU is currently running in some core. In this case,
to avoid unnecessary exits, we try to inject the virtual interrupts using
the paravirtual posted interrupts mechanism. If the VCPU is not running,
we proceed as usual (queue the interrupt).

Signed-off-by: Abel Gordon <abelg@il.ibm.com>