Add virtqueue polling mode to vhost
When vhost is waiting for buffers from the guest driver (e.g., more packets
to send in vhost-net's transmit queue), it normally goes to sleep and waits
for the guest to "kick" it. This kick involves a PIO in the guest, an
therefore an exit (and possibly userspace involvement in translating this PIO
exit into a file descriptor event), all of which hurts performance.
If we can dedicate a core to vhost, we can have it continuously poll the
virtqueues for new buffers, and avoid asking the guest to kick us.

Polling on all virtqueues happens on the same worker thread, in round-robin
fashion. Thanks to the previous patch, the virtqueues of multiple VMs may be
polled on the same worker thread, which allows dedicating only one core to
servicing the I/O from multiple vcpus.

When polling is active for one of the virtqueues, the guest is asked to
disable notification (kicks), and the worker thread continuously checks for
new buffers. When it does discover new buffers, it simulates a "kick" by
invking the  underlying backend driver (such as vhost-net), which thinks it
got a real kick from the guest, and acts accordingly. If the underlying
driver asks not to be kicked, we disable polling on this virtqueue.

In this version, we start polling on a virtqueue when we notice it has
work to do. Polling on this virtqueue is later disabled after 3 seconds of
polling turned out no new work, as in this case we are better off returning
to the exit-based notification mechanism. The default timeout of 3 seconds
can be changed with the "poll_stop_idle" kernel module parameter.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>