Share vhost thread between mutiple devices
Vanilla vhost creates a worker thread and worker queue per virtio device.
This patch creates a worker thread and worker queue shared accross multiple
virtio devices (VMs). The number of maximum virtio devices per worker can be
specified using a kernel module parameter.

Every time a new virtio devices is created we search for a running worker
thread that is serving less than the maximum number of devices. If we find
such worker thread, we bind the new virtio device to this thread. Otherwise,
we create a new worker thread to serve the new virtio devices and more
devices that may be created in the future.

Currently, once a device is binded to a specicific worker thread it can not be
migrated to other worker thread. In the future, we should improve the
mechanism to balance the devices accross threads (migrate a device to a
different worker thread) and to create/destroy threads dynamically during
runtime based on the I/O activity.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>