Do not add a work item if a queue is being polled
vhost stops processing a virqueue once it consumed all the descriptors (processed
all the pending data). To avoid starvation, vhost also limits the amount of data
that can be processed for a virqueue. When the limit is reached, vhost stops
processing the current virtqueue and switches to other virtqueue. vhost may
not receive further notifications for the queue that was limited. Thus, to
ensure that pending data on this queue will be processed later, vhost
adds a new item to the work queue.

If queue is being polled by our mechanism, we don't need to add a new work
item when we limit the queue. We just switch queues and continue polling
as usual.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>