ELI guest kernel module
With ELI, the guest is run using a "shadow IDT" instead of the guest's
requested IDT. This shadow IDT is built by KVM in a way that causes exits
for some interrupts while running the guest's normal handlers on others.

The processor requires that the IDT location be given using a
virtual address, thus, the shadow IDT must always be mapped in the guest
address space. To solve this issue,  this patch introduces a new kernel
module. When this module is loaded, we allocate a page (in the guest)
and pass the corresponding address (GVA) to KVM. ELI can then use
this address to prepare the shadow IDT.

Signed-off-by: Abel Gordon <abelg@il.ibm.com>